lib/fs: Fix and update error about inotify watch limit (fixes #4833)
